feat: Restructure tests
