+++ tests/unit/db/models/test_party.py
